requirements.txt
APScheduler==3.0.0
docker-py==0.4.0
https://api.github.com/repos/totem/yoda-py/tarball/master